requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
idna==3.4
numpy==1.23.3
pandas==1.5.0
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
typing-extensions==4.3.0
urllib3==1.26.12
